requirements.txt
asgiref==3.4.1
Django==3.2.6
django-mssql-backend==2.8.1
pyodbc==4.0.32
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.10.0.1
djangorestframework